+++ xmlunit-matchers/src/test/java/org/xmlunit/matchers/HasXPathMatcherTest.java
@@ -3,13 +3,16 @@ package org.xmlunit.matchers;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.Mockito;
 import org.w3c.dom.Element;
 import org.xml.sax.InputSource;
+import org.xmlunit.XMLUnitException;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
 import java.util.HashMap;
 
 import static org.hamcrest.CoreMatchers.both;
@@ -138,5 +141,27 @@ public class HasXPathMatcherTest {
         String xml = "<a><b attr=\"abc\"></b></a>";
         assertThat(xml, both(not(isEmptyString()))
                    .and(hasXPath("//a/b/@attr")));
+    }
+
+    @Test
+    public void usesDocumentBuilderFactory() throws Exception {
+        DocumentBuilderFactory dFac = Mockito.mock(DocumentBuilderFactory.class);
+        DocumentBuilder b = Mockito.mock(DocumentBuilder.class);
+        Mockito.when(dFac.newDocumentBuilder()).thenReturn(b);
+        Mockito.doThrow(new IOException())
+            .when(b).parse(Mockito.any(InputSource.class));
+
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<fruits>" +
+                    "<fruit name=\"apple\"/>" +
+                    "<fruit name=\"orange\"/>" +
+                    "<fruit name=\"banana\"/>" +
+                "</fruits>";
+        try {
+            assertThat(xml, hasXPath("//fruits/fruit").withDocumentBuilderFactory(dFac));
+            Assert.fail("Expected exception");
+        } catch (XMLUnitException ex) {
+            Mockito.verify(b).parse(Mockito.any(InputSource.class));
+        }
     }
 }